initial conversion to switch mode power supply
linear regulators have fundamental limits (power dissipations in the
transistor), so switch (heh) to a switcher

also, add an inrush current limiting NTC resistor, to eliminate the
frequent fuse blows on startup i'm seeing with the current design

also, decrease the width of power tracks. the previous was massively
overkill.

this has bunch of silkscreen-on-copper drc violations, because the
layout is not finalized yet